refine the folder structure
