+++ ui-primitives/form/src/description.tsx
@@ -1,0 +1,15 @@
+import * as React from 'react'
+import { useFormField } from './useFormField'
+
+const Description = React.forwardRef<
+  HTMLParagraphElement,
+  React.HTMLAttributes<HTMLParagraphElement>
+>(({ ...props }, ref) => {
+  const { formDescriptionId } = useFormField()
+
+  return <p ref={ref} id={formDescriptionId} {...props} />
+})
+
+Description.displayName = 'Description'
+
+export { Description }
