+++ ui-primitives/form/src/index.ts
@@ -1,5 +1,9 @@
-import * as Form from './form'
-
-export { Provider, Item, Label, Control, Description, Message, Field } from './form'
-
-export default Form
+export { Provider } from './provider'
+export { Control } from './control'
+export { Description } from './description'
+export { Field } from './field'
+export { Item } from './item'
+export { Label } from './label'
+export { Message } from './message'
+export { Root } from './root'
+export { useFormField } from './useFormField'
