+++ ui-primitives/form/src/provider.tsx
@@ -1,0 +1,5 @@
+import { FormProvider } from 'react-hook-form'
+
+const Provider = FormProvider
+
+export { Provider }
