+++ ui-primitives/form/src/root.tsx
@@ -1,0 +1,11 @@
+const Root = (props: React.ComponentPropsWithRef<'form'>) => {
+  const { children, ref, ...rest } = props
+
+  return (
+    <form ref={ref} {...rest}>
+      {children}
+    </form>
+  )
+}
+
+export { Root }
